fix: fixing tests
